New async child (#567)
This PR is meant to extend the current AsyncChild to handle multiple
AsyncValues simultaneously to avoid having to use AsyncChild in
AsyncChild.
For example with Phonebook we need to fetch associations and
associationKinds and both are mixed up, thus we can't show any of them
separately and we need to wait for both to be available.
With this one widget would be sufficient.