Remember what store the user selected last (#595)
Feedback: several people scanned several times for the wrong store
because the default one was always the same.

Now we put in the storage the ID of the last store selected by the user,
so that when s/he comes back on MyECL during an event on behalf of one
club, the virtual money goes the right store.